backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
cloud-sql-python-connector[asyncpg]==1.12.1
google-cloud-secret-manager==2.20.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
zeep==4.3.1
firebase-admin==6.5.0
pydantic==2.7.4
